+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 
 [Unreleased]
 
+### Changed
+- Directory listings on the console port now return an HTTP 404 Not Found error.
+
 ## [3.41.0] - 2026-09-18
 ### Added
 - Build with Go 1.27.1.